Model/requirements.txt
flask
python-dotenv
langchain-groq
langchain-huggingface
langchain-community
gunicorn
pinecone-client
sentence-transformers
torch